Moved program selection to separate component
