+++ client/student/components/program/Activity/index.vue
