+++ client/student/components/program/Courseware/ActivityCard.vue
